XHTML2CSV.sh
This is the bash script which downloads the webpage for the weather of seven different cities and uses tagsoup to convert the html to xhtml. Once all the pages are converted to xhtml then the script sends each xhtml file to a python file that takes relevant info from each webpage to be made into a csv file. Finally it repeats every hour